Split redshift sql and cluster objects (#20276)
The first redshift hook was for managing the cluster itself.  Later a hook for _using_ the cluster (e.g. running sql statements) was added to the same module.  Better to separate these into distinct modules redshift_sql and redshift_cluster.  Here we split the redshift modules for operators, hooks, and sensors.
