+++ airflow/contrib/sensors/aws_redshift_cluster_sensor.py
@@ -16,14 +16,16 @@
 # specific language governing permissions and limitations
 # under the License.
 
-"""This module is deprecated. Please use :mod:`airflow.providers.amazon.aws.sensors.redshift`."""
+"""This module is deprecated. Please use :mod:`airflow.providers.amazon.aws.sensors.redshift_cluster`."""
 
 import warnings
 
-from airflow.providers.amazon.aws.sensors.redshift import AwsRedshiftClusterSensor  # noqa
+from airflow.providers.amazon.aws.sensors.redshift_cluster import AwsRedshiftClusterSensor
 
 warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.redshift`.",
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.redshift_cluster`.",
     DeprecationWarning,
     stacklevel=2,
 )
+
+__all__ = ["AwsRedshiftClusterSensor"]
