+++ airflow/providers/amazon/aws/sensors/redshift.py
@@ -15,46 +15,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-from typing import Optional
+import warnings
 
-from airflow.providers.amazon.aws.hooks.redshift import RedshiftHook
-from airflow.sensors.base import BaseSensorOperator
+from airflow.providers.amazon.aws.sensors.redshift_cluster import AwsRedshiftClusterSensor
 
+warnings.warn(
+    "This module is deprecated. Please use `airflow.sensors.redshift_cluster`.",
+    DeprecationWarning,
+    stacklevel=2,
+)
 
-class AwsRedshiftClusterSensor(BaseSensorOperator):
-    """
-    Waits for a Redshift cluster to reach a specific status.
-
-    :param cluster_identifier: The identifier for the cluster being pinged.
-    :type cluster_identifier: str
-    :param target_status: The cluster status desired.
-    :type target_status: str
-    """
-
-    template_fields = ('cluster_identifier', 'target_status')
-
-    def __init__(
-        self,
-        *,
-        cluster_identifier: str,
-        target_status: str = 'available',
-        aws_conn_id: str = 'aws_default',
-        **kwargs,
-    ):
-        super().__init__(**kwargs)
-        self.cluster_identifier = cluster_identifier
-        self.target_status = target_status
-        self.aws_conn_id = aws_conn_id
-        self.hook: Optional[RedshiftHook] = None
-
-    def poke(self, context):
-        self.log.info('Poking for status : %s\nfor cluster %s', self.target_status, self.cluster_identifier)
-        return self.get_hook().cluster_status(self.cluster_identifier) == self.target_status
-
-    def get_hook(self) -> RedshiftHook:
-        """Create and return a RedshiftHook"""
-        if self.hook:
-            return self.hook
-
-        self.hook = RedshiftHook(aws_conn_id=self.aws_conn_id)
-        return self.hook
+__all__ = ["AwsRedshiftClusterSensor"]
