+++ tests/providers/amazon/aws/hooks/test_redshift_cluster.py
@@ -1,4 +1,3 @@
-#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -15,18 +14,13 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-#
 
-import json
 import unittest
-from unittest import mock
 
 import boto3
-from parameterized import parameterized
 
-from airflow.models import Connection
 from airflow.providers.amazon.aws.hooks.base_aws import AwsBaseHook
-from airflow.providers.amazon.aws.hooks.redshift import RedshiftHook, RedshiftSQLHook
+from airflow.providers.amazon.aws.hooks.redshift_cluster import RedshiftHook
 
 try:
     from moto import mock_redshift
@@ -107,67 +101,3 @@ class TestRedshiftHook(unittest.TestCase):
         hook = RedshiftHook(aws_conn_id='aws_default')
         status = hook.cluster_status('test_cluster')
         assert status == 'available'
-
-
-class TestRedshiftSQLHookConn(unittest.TestCase):
-    def setUp(self):
-        super().setUp()
-
-        self.connection = Connection(
-            conn_type='redshift', login='login', password='password', host='host', port=5439, schema="dev"
-        )
-
-        self.db_hook = RedshiftSQLHook()
-        self.db_hook.get_connection = mock.Mock()
-        self.db_hook.get_connection.return_value = self.connection
-
-    def test_get_uri(self):
-        expected = 'redshift+redshift_connector://login:password@host:5439/dev'
-        x = self.db_hook.get_uri()
-        assert x == expected
-
-    @mock.patch('airflow.providers.amazon.aws.hooks.redshift.redshift_connector.connect')
-    def test_get_conn(self, mock_connect):
-        self.db_hook.get_conn()
-        mock_connect.assert_called_once_with(
-            user='login', password='password', host='host', port=5439, database='dev'
-        )
-
-    @mock.patch('airflow.providers.amazon.aws.hooks.redshift.redshift_connector.connect')
-    def test_get_conn_extra(self, mock_connect):
-        self.connection.extra = json.dumps(
-            {
-                "iam": True,
-                "cluster_identifier": "my-test-cluster",
-                "profile": "default",
-            }
-        )
-        self.db_hook.get_conn()
-        mock_connect.assert_called_once_with(
-            user='login',
-            password='password',
-            host='host',
-            port=5439,
-            cluster_identifier="my-test-cluster",
-            profile="default",
-            database='dev',
-            iam=True,
-        )
-
-    @parameterized.expand(
-        [
-            ({}, {}, {}),
-            ({"login": "test"}, {}, {"user": "test"}),
-            ({}, {"user": "test"}, {"user": "test"}),
-            ({"login": "original"}, {"user": "overridden"}, {"user": "overridden"}),
-            ({"login": "test1"}, {"password": "test2"}, {"user": "test1", "password": "test2"}),
-        ],
-    )
-    @mock.patch('airflow.providers.amazon.aws.hooks.redshift.redshift_connector.connect')
-    def test_get_conn_overrides_correctly(self, conn_params, conn_extra, expected_call_args, mock_connect):
-        with mock.patch(
-            'airflow.providers.amazon.aws.hooks.redshift.RedshiftSQLHook.conn',
-            Connection(conn_type='redshift', extra=conn_extra, **conn_params),
-        ):
-            self.db_hook.get_conn()
-            mock_connect.assert_called_once_with(**expected_call_args)
