+++ tests/providers/amazon/aws/operators/test_redshift_cluster.py
@@ -1,4 +1,3 @@
-#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -16,36 +15,12 @@
 # specific language governing permissions and limitations
 # under the License.
 
-import unittest
 from unittest import mock
-from unittest.mock import MagicMock
 
-from parameterized import parameterized
-
-from airflow.providers.amazon.aws.operators.redshift import (
+from airflow.providers.amazon.aws.operators.redshift_cluster import (
     RedshiftPauseClusterOperator,
     RedshiftResumeClusterOperator,
-    RedshiftSQLOperator,
 )
-
-
-class TestRedshiftSQLOperator(unittest.TestCase):
-    @parameterized.expand([(True, ('a', 'b')), (False, ('c', 'd'))])
-    @mock.patch("airflow.providers.amazon.aws.operators.redshift.RedshiftSQLOperator.get_hook")
-    def test_redshift_operator(self, test_autocommit, test_parameters, mock_get_hook):
-        hook = MagicMock()
-        mock_run = hook.run
-        mock_get_hook.return_value = hook
-        sql = MagicMock()
-        operator = RedshiftSQLOperator(
-            task_id='test', sql=sql, autocommit=test_autocommit, parameters=test_parameters
-        )
-        operator.execute(None)
-        mock_run.assert_called_once_with(
-            sql,
-            autocommit=test_autocommit,
-            parameters=test_parameters,
-        )
 
 
 class TestResumeClusterOperator:
