+++ tests/providers/amazon/aws/operators/test_redshift_sql.py
@@ -1,4 +1,3 @@
-#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -15,19 +14,30 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-import warnings
 
-from airflow.providers.amazon.aws.operators.redshift_cluster import (
-    RedshiftPauseClusterOperator,
-    RedshiftResumeClusterOperator,
-)
+import unittest
+from unittest import mock
+from unittest.mock import MagicMock
+
+from parameterized import parameterized
+
 from airflow.providers.amazon.aws.operators.redshift_sql import RedshiftSQLOperator
 
-warnings.warn(
-    "This module is deprecated. Please use `airflow.operators.redshift_sql` "
-    "or `airflow.operators.redshift_cluster` as appropriate.",
-    DeprecationWarning,
-    stacklevel=2,
-)
 
-__all__ = ["RedshiftSQLOperator", "RedshiftPauseClusterOperator", "RedshiftResumeClusterOperator"]
+class TestRedshiftSQLOperator(unittest.TestCase):
+    @parameterized.expand([(True, ('a', 'b')), (False, ('c', 'd'))])
+    @mock.patch("airflow.providers.amazon.aws.operators.redshift_sql.RedshiftSQLOperator.get_hook")
+    def test_redshift_operator(self, test_autocommit, test_parameters, mock_get_hook):
+        hook = MagicMock()
+        mock_run = hook.run
+        mock_get_hook.return_value = hook
+        sql = MagicMock()
+        operator = RedshiftSQLOperator(
+            task_id='test', sql=sql, autocommit=test_autocommit, parameters=test_parameters
+        )
+        operator.execute(None)
+        mock_run.assert_called_once_with(
+            sql,
+            autocommit=test_autocommit,
+            parameters=test_parameters,
+        )
